+++ Funcky.Xunit/FunctionalAssert/IsSome.cs
@@ -50,10 +50,7 @@ namespace Funcky
                     none: static () => throw new AssertActualExpectedException(
                         expected: "Some(...)",
                         actual: "None",
-                        userMessage: $"{nameof(FunctionalAssert)}.{nameof(IsSome)}() Failure",
-                        expectedTitle: null, // The other constructor overload is missing in 2.4.2-pre.12. See https://github.com/xunit/xunit/issues/2449
-                        actualTitle: null,
-                        innerException: null));
+                        userMessage: $"{nameof(FunctionalAssert)}.{nameof(IsSome)}() Failure"));
             }
             catch (AssertActualExpectedException exception)
             {